test: move timelock tests to new file
